ignore

+++ APCS-A-2020/src/Unit2/Line.java
@@ -7,9 +7,9 @@ package Unit2;
 
 public class Line
 {
-	public static double getSlope( double x1, double y1, double x2, double y2 )
+	public static double getSlope( int x1, int y1, int x2, int y2 )
 	{
-		return ((y1 - y2) / (x1 - x2));
+		return  ( (double) (y1 - y2) / (x1 - x2));
 	}
 
 }
